+++ test/test.test.tsx
@@ -1,5 +1,0 @@
-
-
-it('should test true', () => {
-  expect(true).toEqual(true);
-});